+++ src/rules/selector-attribute-operator-whitelist/index.js
@@ -13,7 +13,7 @@ export const messages = ruleMessages(ruleName, {
   rejected: (operator) => `Unexpected operator "${operator}"`,
 })
 
-export default function (whitelistInput) {
+function rule(whitelistInput) {
   const whitelist = [].concat(whitelistInput)
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
@@ -46,3 +46,7 @@ export default function (whitelistInput) {
     })
   }
 }
+
+rule.primaryOptionArray = true
+
+export default rule